Update the "What is Jim" section
Jim was really only branded as it's own project to appease the powers that be, however, I wasn't against the idea initially. But after discovering other technologies that were already more or less in line with what Jim wanted to be, I decided there was no point in expanding Jim to anything other than an FFI extension for Elm.